Downgrade to 0.27 because the library is not upgraded to comply with RN0.30 yet. iOS example working. Will revisit if the other PR is merged.
