+++ example/android/app/build.gradle
@@ -55,13 +55,7 @@ import com.android.build.OutputFile
  *   // date; if you have any other folders that you want to ignore for performance reasons (gradle
  *   // indexes the entire tree), add them here. Alternatively, if you have JS files in android/
  *   // for example, you might want to remove it from here.
- *   inputExcludes: ["android/**", "ios/**"],
- *
- *   // override which node gets called and with what additional arguments
- *   nodeExecutableAndArgs: ["node"]
- *
- *   // supply additional arguments to the packager
- *   extraPackagerArgs: []
+ *   inputExcludes: ["android/**", "ios/**"]
  * ]
  */
 
@@ -110,10 +104,6 @@ android {
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
     }
-    packagingOptions {
-        exclude 'META-INF/LICENSE'
-        exclude 'META-INF/NOTICE'
-    }
     // applicationVariants are e.g. debug, release
     applicationVariants.all { variant ->
         variant.outputs.each { output ->
@@ -133,8 +123,6 @@ dependencies {
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
-    compile group: 'com.zopim.android', name: 'sdk', version: '1.2.0.1'
-    compile project(':react-native-zendesk-chat')
 }
 
 // Run this once to be able to run the application with BUCK
@@ -143,4 +131,3 @@ task copyDownloadableDepsToLibs(type: Copy) {
   from configurations.compile
   into 'libs'
 }
-
