+++ example/android/app/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
         android:targetSdkVersion="22" />
 
     <application
-      android:name=".MainApplication"
       android:allowBackup="true"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
