+++ example/android/app/src/main/java/com/example/MainActivity.java
@@ -1,10 +1,11 @@
 package com.example;
 
-import android.os.Bundle;
-
 import com.facebook.react.ReactActivity;
+import com.facebook.react.ReactPackage;
+import com.facebook.react.shell.MainReactPackage;
 
-import com.zopim.android.sdk.api.ZopimChat;
+import java.util.Arrays;
+import java.util.List;
 
 public class MainActivity extends ReactActivity {
 
@@ -17,10 +18,23 @@ public class MainActivity extends ReactActivity {
         return "example";
     }
 
+    /**
+     * Returns whether dev mode should be enabled.
+     * This enables e.g. the dev menu.
+     */
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        // Init ZopimChat
-        ZopimChat.init("3q3y4vnmd334dHpK6EvVALjPtSLWzNcb");
+    protected boolean getUseDeveloperSupport() {
+        return BuildConfig.DEBUG;
+    }
+
+    /**
+     * A list of packages used by the app. If the app uses additional views
+     * or modules besides the default ones, add more packages here.
+     */
+    @Override
+    protected List<ReactPackage> getPackages() {
+        return Arrays.<ReactPackage>asList(
+            new MainReactPackage()
+        );
     }
 }
