+++ example/android/build.gradle
@@ -18,8 +18,7 @@ allprojects {
         jcenter()
         maven {
             // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
-            url "$rootDir/../node_modules/react-native/android"
+            url "$projectDir/../../node_modules/react-native/android"
         }
-        maven { url 'https://zendesk.artifactoryonline.com/zendesk/repo' }
     }
 }
