+++ example/android/settings.gradle
@@ -1,6 +1,3 @@
 rootProject.name = 'example'
 
 include ':app'
-
-include ':react-native-zendesk-chat'
-project(':react-native-zendesk-chat').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-zendesk-chat/android')
